# speed up face mask render using cython