--- a/addition_module/face_mask_adding/FMA-3D/.gitignore
+++ b/addition_module/face_mask_adding/FMA-3D/.gitignore
@@ -1,0 +1,6 @@
+__pycache__
+test1_mask1.jpg
+build
+*.c
+*.html
+*.so