--- a/addition_module/face_mask_adding/FMA-3D/utils/cython/README.md
+++ b/addition_module/face_mask_adding/FMA-3D/utils/cython/README.md
@@ -1,13 +1,17 @@
 ## What's this
+
 this is a Cython implementation to speed up adding mask rendering process
 
 ## Usage
+
 just run `python setup.py` to generate a share file
 
-## Compare 
-As an example, for **add_mask_one.py** on my own computer, cumtime in `face_masker.add_mask_one` is down from to 0.249s compare to origin method.
+## Compare
+
+As an example, for **add_mask_one.py** on my own computer, cumtime in **face_masker.add_mask_one** is down from10.885s to 0.249s compare to origin method
 
 ## Other
-Author: [cbwces](https://github.com/cbwces)
-Mail: sknyqbcbw@gmail.com
 
+Author: [cbwces](https://github.com/cbwces)
+
+Mail: sknyqbcbw@gmail.com