--- a/addition_module/face_mask_adding/FMA-3D/utils/cython/setup.py
+++ b/addition_module/face_mask_adding/FMA-3D/utils/cython/setup.py
@@ -1,0 +1,17 @@
+from setuptools import setup, Extension
+from Cython.Build import cythonize
+import numpy
+
+ext_modules = [
+    Extension(
+        "render",
+        ["render.pyx"],
+        extra_compile_args=['-fopenmp'],
+        extra_link_args=['-fopenmp'],
+    )
+]
+
+setup(
+    ext_modules=cythonize(ext_modules),
+    include_dirs=[numpy.get_include()]
+)