--- a/addition_module/face_mask_adding/FMA-3D/utils/cython/setup.py
+++ b/addition_module/face_mask_adding/FMA-3D/utils/cython/setup.py
@@ -14,4 +14,4 @@ ext_modules = [
 setup(
     ext_modules=cythonize(ext_modules),
     include_dirs=[numpy.get_include()]
-)    
+)