add render.pyx and setup.py under `FaceX-Zoo/addition_module/face_mask_adding/FMA-3D/utils/cython` to implement render

**before** python add_mask_one.py  11.24s user 2.37s system 117% cpu 11.593 total
**now** python add_mask_one.py  3.25s user 2.34s system 159% cpu 3.506 total