@luckycaicai 2021-05-27
In face_mask_adding, it made some noise on the image ,when using 'Speed Up' method. More detail in   #78